moving the auto refresh selector to its own file
